requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
databricks-sdk==0.88.0
debugpy==1.8.20
decorator==5.2.1
docker==7.1.0
executing==2.2.1
fastapi==0.129.0
filelock==3.21.2
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
gunicorn==23.0.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.0
kagglehub==1.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mplcyberpunk==0.7.6
mpmath==1.3.0
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.7.0
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.46
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchutils==0.0.4
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.5
zipp==3.23.0
